refactor: Mover enums para o domínio e corrigir pacotes de entidades
- Movidos os enums `EspecialidadeEnum`, `PerfilEnum` e `StatusConsultaEnum` da infraestrutura (`infrastructure.enumerate`) para o domínio (`domain.enumerator`), garantindo melhor separação das regras de negócio.
- Atualizados os imports das entidades para refletir a nova localização dos enums.
- Ajustado o pacote das entidades para `infrastructure.persistence.entity`, organizando melhor a persistência.

Essas mudanças seguem a Arquitetura Limpa, mantendo o domínio independente de implementações específicas da infraestrutura.
